copy to root material_3_demo
