+++ material_3_demo/pubspec.yaml
@@ -22,6 +22,8 @@ dev_dependencies:
     sdk: flutter
 
   flutter_lints: ^2.0.1
+  integration_test:
+    sdk: flutter
 
 flutter:
   uses-material-design: true
