+++ material_3_demo/test/component_screen_test.dart
@@ -11,7 +11,7 @@ import 'package:material_3_demo/main.dart';
 void main() {
   testWidgets('Default main page shows all M3 components', (tester) async {
     widgetSetup(tester, 800, windowHeight: 7000);
-    await tester.pumpWidget(const MaterialApp(home: Material3Demo()));
+    await tester.pumpWidget(const App());
 
     // Elements on the app bar
     expect(find.text('Material 3'), findsOneWidget);
@@ -131,7 +131,7 @@ void main() {
       'NavigationRail doesn\'t show when width value is small than 1000 '
       '(in Portrait mode or narrow screen)', (tester) async {
     widgetSetup(tester, 999, windowHeight: 7000);
-    await tester.pumpWidget(const MaterialApp(home: Material3Demo()));
+    await tester.pumpWidget(const App());
     await tester.pumpAndSettle();
 
     // When screen width is less than 1000, NavigationBar will show. At the same
@@ -152,7 +152,7 @@ void main() {
       'NavigationRail shows when width value is greater than or equal '
       'to 1000 (in Landscape mode or wider screen)', (tester) async {
     widgetSetup(tester, 1001, windowHeight: 3000);
-    await tester.pumpWidget(const MaterialApp(home: Material3Demo()));
+    await tester.pumpWidget(const App());
     await tester.pumpAndSettle();
 
     // When screen width is greater than or equal to 1000, NavigationRail will show.
@@ -178,7 +178,7 @@ void main() {
       'Material version switches between Material3 and Material2 when'
       'the version icon is clicked', (tester) async {
     widgetSetup(tester, 450, windowHeight: 7000);
-    await tester.pumpWidget(const MaterialApp(home: Material3Demo()));
+    await tester.pumpWidget(const App());
     BuildContext defaultElevatedButton =
         tester.firstElement(find.byType(ElevatedButton));
     BuildContext defaultIconButton =
@@ -244,7 +244,7 @@ void main() {
   testWidgets(
       'Other screens become Material2 mode after changing mode from '
       'main screen', (tester) async {
-    await tester.pumpWidget(const MaterialApp(home: Material3Demo()));
+    await tester.pumpWidget(const App());
     Finder appbarM2Icon = find.descendant(
         of: find.byType(AppBar),
         matching: find.widgetWithIcon(IconButton, Icons.filter_2));
@@ -279,7 +279,7 @@ void main() {
   testWidgets(
       'Brightness mode switches between dark and light when'
       'the brightness icon is clicked', (tester) async {
-    await tester.pumpWidget(const MaterialApp(home: Material3Demo()));
+    await tester.pumpWidget(const App());
     Finder lightIcon = find.descendant(
         of: find.byType(AppBar),
         matching: find.widgetWithIcon(IconButton, Icons.light_mode_outlined));
@@ -314,7 +314,7 @@ void main() {
       (tester) async {
     Color m3BaseColor = const Color(0xff6750a4);
     await tester.pumpWidget(Container());
-    await tester.pumpWidget(const MaterialApp(home: Material3Demo()));
+    await tester.pumpWidget(const App());
     await tester.pump();
     Finder menuIcon = find.descendant(
         of: find.byType(AppBar),
